/ mount: should be readonly, with --inplace --sparse it will fail
not exactly sure why it fails only when --sparse, but there's no reason
the backup mount needs to be readwrite anyways.  in fact it shouldn't so
this is an accident but helped us to find an underlying bug

possibly it's a race between when the file gets stated for time, and
then it gets modified after that, while the copy is in process.  it
shouldn't happen though since is just a snapshot.  not sure understand
why this is happening but --inplace --sparse can only be used together
new in 3.1.3 and then there's some sparse fixes in master after that but
before the next release.  since it works with readonly, and that's the
extent of our requirement, will leave this alone, but it doesn't make
much sense that it works fine readwrite, but only when it's not on a
snapshot.